requirements.txt
azure-iot-device
azure-iot-hub
azure-iothub-service-client
azure-iothub-device-client
pandas
mysql-connector-python
cryptography
websockets==8.1
asyncio
hbmqtt
